+++ src/Pages/Dashboard/MyReportList/MyReportList.js
@@ -2,7 +2,7 @@ import { useQuery } from '@tanstack/react-query';
 import axios from 'axios';
 import React from 'react';
 import { useContext } from 'react';
-import { toast } from 'react-toastify';
+import swal from 'sweetalert';
 import Loader from '../../../Components/Loader/Loader';
 import { AuthContext } from '../../../Contexts/AuthProvider';
 
@@ -22,18 +22,33 @@ const MyReportList = () => {
 
     //delete order from the server
     const handleDeleteReport = id => {
-        axios.delete(`${process.env.REACT_APP_API_URL}/rportings/${id}`, {
-            headers: {
-                authorization: `bearer ${localStorage.getItem('accessToken')}`
-            }
+        swal({
+            title: "Are you sure?",
+            text: "Once deleted, you will not be able to recover this report!",
+            icon: "warning",
+            buttons: true,
+            dangerMode: true,
         })
-            .then(sellerData => {
-                if (sellerData.data.acknowledged) {
-                    toast.success('Report has been deleted successfully!')
-                    refetch()
+            .then((willDelete) => {
+                if (willDelete) {
+                    axios.delete(`${process.env.REACT_APP_API_URL}/rportings/${id}`, {
+                        headers: {
+                            authorization: `bearer ${localStorage.getItem('accessToken')}`
+                        }
+                    })
+                        .then(sellerData => {
+                            if (sellerData.data.acknowledged) {
+                                refetch()
+                            }
+                        })
+                        .catch(error => console.log(error))
+                    swal('Your report has been deleted successfully!', {
+                        icon: "success",
+                    });
+                } else {
+                    swal("Your report is safe!");
                 }
-            })
-            .catch(error => console.log(error))
+            });
     }
 
     if (isLoading) {
